Update ui folder structure
